examples/langchain/requirements.txt
langchain
langchain-openai
httpx
web3
python-dotenv
